refactor(layout): introduce domain/ subpackage with pure leaves
Step 1 of the domain/ reorganization plan from the post-#1192
audit. Moves the three modules with zero internal LCM imports
into a new domain/ subpackage:

* exceptions.py -> domain/exceptions.py
* resilience.py -> domain/resilience.py
* pin_generator.py -> domain/pin_generator.py

These are the safest leaves: each has no internal imports of its
own, so the move is mechanically pure. All importers across root,
providers/, and tests/ are updated to the new canonical path
rather than introducing back-compat shim re-exports; the auditor
flagged shims as the kind of layering noise this reorganization
exists to remove.

The domain/ package will be the home for LCM-internal logic that
should not be imported through the HA-platform-shim layer. Future
steps will move data types (models, EntryConfig), domain logic
(coordinator, sync, slot coordinator), and the queries/services
split out of helpers.py/data.py.
