+++ custom_components/lock_code_manager/domain/__init__.py
@@ -1,0 +1,1 @@
+"""Domain (LCM-internal) modules. See AGENTS.md for the layering."""
