+++ custom_components/lock_code_manager/domain/exceptions.py
