+++ custom_components/lock_code_manager/providers/akuvox.py
@@ -20,7 +20,7 @@ from typing import Any, Literal
 
 from homeassistant.config_entries import ConfigEntry
 
-from ..exceptions import (
+from ..domain.exceptions import (
     LockCodeManagerProviderError,
     LockDisconnected,
     LockOperationFailed,
