+++ custom_components/lock_code_manager/providers/zigbee2mqtt.py
@@ -19,7 +19,7 @@ from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import callback
 from homeassistant.exceptions import HomeAssistantError
 
-from ..exceptions import LockDisconnected, LockOperationFailed
+from ..domain.exceptions import LockDisconnected, LockOperationFailed
 from ..models import SlotCredential
 from ._base import BaseLock
 from ._util import parse_slot_num
