+++ tests/providers/zwave_js/test_provider.py
@@ -20,7 +20,7 @@ from custom_components.lock_code_manager.const import (
     CONF_SLOTS,
     DOMAIN,
 )
-from custom_components.lock_code_manager.exceptions import LockDisconnected
+from custom_components.lock_code_manager.domain.exceptions import LockDisconnected
 from custom_components.lock_code_manager.models import SlotCredential
 from custom_components.lock_code_manager.providers.zwave_js import ZWaveJSLock
