+++ tests/test_binary_sensor.py
@@ -45,7 +45,7 @@ from custom_components.lock_code_manager.const import (
 from custom_components.lock_code_manager.coordinator import (
     LockUsercodeUpdateCoordinator,
 )
-from custom_components.lock_code_manager.exceptions import (
+from custom_components.lock_code_manager.domain.exceptions import (
     DuplicateCodeError,
     LockCodeManagerError,
 )
