+++ tests/test_coordinator.py
@@ -22,7 +22,7 @@ from custom_components.lock_code_manager.const import (
 from custom_components.lock_code_manager.coordinator import (
     LockUsercodeUpdateCoordinator,
 )
-from custom_components.lock_code_manager.exceptions import LockDisconnected
+from custom_components.lock_code_manager.domain.exceptions import LockDisconnected
 from custom_components.lock_code_manager.models import SlotCredential
 from custom_components.lock_code_manager.providers.virtual import VirtualLock
